+++ renter-api/Datebase/Interfaces/IRentBookRepositoryService.cs
@@ -7,11 +7,11 @@ namespace Database.Interfaces
 {
     public interface IRentBookRepositoryService : IRepositoryService<RentBook>
     {
-        RentBook Rent(int bookId, string userId, TimeSpan time);
-        bool IsBookAvailable(int bookId);
+        RentBook Rent(long bookId, string userId, TimeSpan time);
+        bool IsBookAvailable(long bookId);
         List<Book> GetAvaiableBooksByTitle(string title);
         List<Book> GetAvaiableBooksByIsbn(string isbn);
         List<RentBook> GetUserRentHisotry(string userId);
-        List<RentBook> GetBookRentHisotry(int bookId);
+        List<RentBook> GetBookRentHisotry(long bookId);
     }
 }
